# Conflicts:
#	CMakeLists.txt
#	lib/huji-rich-Elad3DFast/FindVORONOI.cmake
#	vivid

+++ pybind11/include/pybind11/detail/common.h
@@ -688,7 +688,7 @@ template <typename T> struct is_fmt_numeric<T, enable_if_t<std::is_arithmetic<T>
     static constexpr bool value = true;
     static constexpr int index = std::is_same<T, bool>::value ? 0 : 1 + (
         std::is_integral<T>::value ? detail::log2(sizeof(T))*2 + std::is_unsigned<T>::value : 8 + (
-        std::is_same<T, double>::value ? 1 : std::is_same<T, long double>::value ? 2 : 0));
+        std::is_same<T, double>::value ? 1 : std::is_same<T, double>::value ? 2 : 0));
 };
 NAMESPACE_END(detail)
